kvm_monitor_to_csv_xls/requirements.txt
psycopg2-binary
PyYaml
xlwt
requests
